Fix RawSyntax's leading/trailingTrivia computation to return nil if its outermost children don't have leading/trailing trivia
Previously if would skip over outer children that didn't have leading/trailing
trivia until it found one that did. This was causing that trivia to be included
twice when computing the totalLength/byteSize of the node; once in
leading/trailingTrivia and again in contentlength, which only skipped the trivia
of the outmost children.