Introduce SyntaxParser that uses the C parsing library from the toolchain
This replaces the previous mechanism that was invoking swiftc in order to get a syntax tree from a file.
There is also a mechanism added for performing incremental re-parsing.

SwiftSyntax now can parse (and re-parse) swift sources on its own, without needing swiftc or swift-syntax-test.
Due to these changes, lit-test-helper is enhanced to allow it to perform the incremental re-parsing tests without
needing to invoke swift-syntax-test from the swift repository.