Speed-up lookups for incremental re-parsing
Previously the search for nodes to re-use was naively doing a lookup that always started from the top of the tree, every time the parser called back.
The new mechanism takes advantage of the fact that the parser continuously calls back with ascending source positions, not random ones, and keeps
a "cursor" state for re-starting the lookup search from the last node position.

This commit also introduces `SyntaxNode` which is a more efficient representation for a syntax tree node, that `SyntaxParser` uses to efficiently
report which syntax nodes got re-used during incremental re-parsing.

Changes result in improved performance for incremental reparsing. When using the test case from rdar://48511326, performance improves by 321x.