Make Diagnostic initializers public
For tools that use SwiftSyntax, it's quite useful to be able to emit their own diagnostics while reusing the `Diagnostic` type from SwiftSyntax. Both `Diagnostic` and its `Builder` type are `public`, and properties of `Diagnostic` are `public` too, but initializers are implicitly `internal`. This means that instances of `Diagnostic` can't be created by tools using SwiftSyntax. I hope this can be fixed with this simple change.