Explicitly add rpath for lib_InternalSwiftSyntaxParser.dylib
Previously, we were relying on SwiftPM adding the toolchain’s stdlib to the rpaths and we were finding `lib_InternalSwiftSyntaxParser.dylib` inside the stdlib directory. This behavior was removed in apple/swift-package-manager#4208.

Explicitly add the directory that contains `lib_InternalSwiftSyntaxParser.dylib` as an rpath to SwiftSyntaxParser when compiling SwiftSyntax using build-script.py.

The major downsides of this approach are
- You can no longer run SwiftSyntaxParser tests for SwiftSyntax in Xcode with an open source toolchain
  - Maybe this isn’t too bad because it only affects SwiftSyntaxParser tests and those can be run from the command line
- All adopters of SwiftSyntax also need to explicitly specify the `lib_InternalSwiftSyntaxParser.dylib` rpath
  - Maybe this isn’t too bad because on Linux `lib_InternalSwiftSyntaxParser.dylib` is in the standard search paths and on macOS we start shipping `lib_InternalSwiftSyntaxParser.dylib` as a binary dependency.