Add an API to compare two separate syntax trees for equivalence
Given two syntax trees, add an API to find the first difference between
them. Use this to provide an XCTest assertion that two trees are equal,
which dumps out both with a marker for the different node otherwise.

Modifies `Format` to allow removing the builder's auto-formatting so
that the trivia in the expected case is more obvious.

Adds a new `SwiftSyntaxSupport` target and moves the various test
methods into it.