Introduce a new syntax kind to represent missing base nodes
In the future, we only want to attach a source presence to tokens, not to nodes. All concreate nodes which are missing can be represented by creating the node and marking all tokens as missing, but if a syntax node carries a child, that has a base kind (like `Decl`), we can’t decide which concrete node to instantiate. Introduce `MissingDecl` etc. node for this purpose.

While at it, also remove the base kinds from `SyntaxKind` because we should never instantiate nodes of these base kinds.

rdar://97908258
rdar://97775360