Make garbage parameters in SyntaxFactory unlabeled
If the `garbage` parameters in the SyntaxFactory are labelled, it’s not possible to specify e.g. the second garbage node but leave the others to their default (https://github.com/apple/swift/issues/60274).

We could either give these `garbage` parameters distinct names or make them unlabeled to work around that issue. I chose to go for the unlabeled approach because it should be clear that these parameters contain garbage because of their type.