Operator-precedence parsing for sequence expressions
Implement the sequence "fold" operation that takes a sequence
expression (i.e., something like a + b * c) and turns it into a
structured syntax tree that reflects the relative precedence of the
operators involved (here, * has higher precedence than +). The core of
this algorithm is a direct port of the C++ implementation
of operator-precedence parsing. There are several pieces to it:

* PrecedenceGroup and Operator, which are semantic representations of
precedence groups and operator declarations.
* PrecedenceGraph, which keeps track of a set of precedence groups and
their relationships to each other, and is able to answer questions
like "how to the precedences of two different groups compare?".
* Syntactic -> semantic translation to go from
PrecedenceGroupDeclSyntax and OperatorDeclSyntax nodes to
PrecedenceGroup and Operator, by walking the child syntax.
* OperatorPrecedence, which contains a PrecedenceGraph that can be
created by processing a source file to pick out the various operator
and precedence-group declarations. It provides the core "fold"
operation that turns a SequenceExprSyntax node into one that is
fully-structured to show the order of operations.

The parsing infrastructure and basic algorithm work (modulo lots of
testing), but there are a lot of missing pieces:

* Right now we either throw an error or outright ignore pretty much
any problem that shows up. Neither of those is an acceptable answer,
so we'll need to decide how to
* There are default precedence groups for certain expression kinds
(e.g., '=', 'as?', '->') that we will need to implement by looking at
the syntax node kind for the "operator" and handling each of the
cases. This is easy but tedious to do/test.
* There is no syntax node for "infix binary expression" in the
grammar, so we cannot actually represent the result the way we want
to. For testing purposes now, we create a new TupleExprSyntax instead,
so we can see the full parenthesization in the output.
* We do not have primitives for merging precedence graphs when
combining multiple source files, modules, etc. We'll need to define
those rules.
* We don't make any attempt at caching repeated queries, which could
matter later on.
* We probably want to provide some pre-canned OperatorPrecedence
instances for, e.g., the logical operators (&&, ||, !) used when
evaluating preprocessor conditionals (#if x || y) and possibly the
standard library itself.

The biggest question of all is whether any of this even makes sense to
do as part of the syntax library, or whether it belongs with the main
Swift compiler.