Remove the old (non-result-builder) SyntaxBuilder
We’ve got by far to many ways to generate SwiftSyntax nodes. Remove the old SyntaxBuilder because, AFAICT, it’s not really used much.

This PR migrates tests from the older builder to SyntaxFactory. I’ll likely remove SyntaxFactory in a follow-up PR in favor of initializers or static methods on the syntax types you want to create.

rdar://98318651