Introduce RawSyntaxData.ParsedToken
A new RawSyntax data variation. This is intended to be used for parsed
tokens. Actual usage will be in follow-ups.

In this variation, unlike the existing `MaterializedToken`, leading/trailing
trivia are stored in a SyntaxText form, and are lazily materialized by a
function registered in `SyntaxArena`. Since the parser doesn't need to
materialize them while parsing, it'd gain the "parsing" performance boost,
in exchange for a performance reduction when accessing trivia.