Make `presence` a token-only property
It doesn’t really make sense to talk about presence for layout nodes. Are they missing if all of their tokens are missing? What if all tokens are missing but one has trivia attached to it?

To simplify the semantics, define `presence` for tokens only.

This slightly changes the semantics because layout nodes without any present children are now walked in the `sourceAccurate` view, but I think that’s a good change.

rdar://98818987