Add infrastructure to make it easy to test the result of sequence folding.
Introduce an assertion operation `assertExpectedFold` that takes a string
containing the source code to be folded along with a fully-parenthesized
version of the same source code. The former will go through the sequence
folding operation, while the latter will go through a separate pass that
identifiers parenthesized sequence expressions for a binary operator and
map them to infix expression syntax. If it fails, we'll get a syntax tree
diff.

Example usage:

  let opPrecedence = OperatorPrecedence.logicalOperators
  try opPrecedence.assertExpectedFold("x || y && w", "(x || (y && w))")