Get rid of `XXXSyntax.Cursor` type
They are internal types which served as layout 'subscript' keys e.g.
`layout[Cursor.propertyName]`. That looks nice, but in reality, the
name is usually written very close:

  var propertyName: PropertyType {
    get { layout[Cursor.propertyName] }
  }

So it didn't give us much benefit. In the contrary, it hid the "index"
information which might be useful when investigating issues. So just get
rid of them for now.

Since they are just internal types. So no external API changes.