Use a dedicated `Token` type in SwiftSyntaxBuilder
At the moment, tokens were a little bit of an oddity in SwiftSyntaxBuilder: All nodes were represented by dedicated types in SwiftSyntaxBuilder but tokens were represented by the `TokenSyntax` from the SwiftSyntax module. This has lead to a few problems:
- We needed to add extensions on `TokenSyntax` to add convienience properties that are only designed for SwiftSyntaxBuilder
- `TokenSyntax` was not `SyntaxBuildable` and thus couldn’t be used as elements in unexpected node collections

In general, it seems cleaner if we just add a custom `Token` type in SwiftSyntaxBuilder.

rdar://99030016