Parse Much Harder for a Right Bracket in Capture Specifiers
All the lookahead we did before this promised us a right bracket here, but the precedence of right brackets is too weak to overcome keywords. The attached test case shows that we fall all the way through closure signature parsing and nothing picks up the slack and eats the remaining parts of the signature.

Instead, do `expect`'s job for it and try to munge through to the right bracket ourselves.

Fixes #721

rdar://99657721