Emulate the Structure of Pattern Binding Conditions
The expression parser has to go out of its way to parse all the
way down to primary expressions before realizing that the expression
pattern it was after isn't actually an expression pattern. This involves
setting a bit of ambient state, that we have to plumb down expression parsing.

This is pretty gross, but it emulates what the legacy parser was doing here.

Fixes #724

rdar://99669036