Provide API for working with syntax text that includes malformed UTF-8.
The `description` operation on a syntax node cannot represent invalid
UTF-8. Document this limitation, and introduce `syntaxTextBytes`, which
produces an array of bytes that accurately represents the source, even
in the presence of malformed UTF-8.

Switch the parser consistency check's round-tripping validation over
to use `syntaxTextBytes`, so that we correctly verify round-tripping
of malformed UTF-8.