Forward Consumption Handles to Decl Productions
canRecoverTo and expect have different precedence models, so it's
not enough to use one on one side and pair it with the other. In this
case, it's because strong structure markers like braces cannot be
overcome by the latter, but _can_ be overcome by the former. So we
could potentially not make progress in some cases.

Fixes #770

rdar://99766200