Enable the creation of syntax nodes from a string interpolation
Make the main categories of syntax nodes (`DeclSyntax`,
`ExprSyntax`, `StmtSyntax`, `TypeSyntax`, and `PatternSyntax`)
conform to `ExpressiblyByStringInterpolation`. The string literal
portions are source text, and one can interpolate both additional
source text and syntax nodes into that source text. The result
will be parsed by the parser to produce a node of the given
kind. For example, we can produce a `TypeSyntax` like this:

    let tupleSyntax: TypeSyntax = "(Int, name: String)"

And then use that in another production, for example to create a
function declaration:

    let fnSyntax: DeclSyntax =
       "func \(name)(String) async throws -> \(tupleSyntax)"

This approach is intended to simplify code generation, by
allowing code to generate Swift syntax trees in the most natural
way possible---by writing the Swift code that generates them.