Translate test case from swift/test/Parse to SwiftParserTest
This translates the test cases from swift/test/Parse into XCTests that can be run in the swift-syntax repository.

All previously expected errors, warnings and notes are imported as TODOs, while the expected diagnostics are filled with those the parser currently produces.

The idea is that this commit just performs a mechanicaly 1:1 translation. Upcoming commits will need to address the TODOs by

removing them (e.g. because the diagnostics are semantic and should be reported by ASTGen in the new scheme)
matching them up with diagnostics that are already produced by the new parser
Adjusting the new parser to produce the previously expected diagnostics.
Powering through these TODOs will be quite a bit of effort but it’s something that I think we need to do…