Allow combining SwiftParser and SwiftSyntaxBuilder to generate source code
This is a proof-of-concept how the parser can be used to generate SwiftSyntax trees for statically known boilerplate while using SwiftSyntaxBuilder for repetitive or dynamic constructs.

Once this is merged, I will try to migrate generate-swiftsyntaxbuilder to this scheme to make it easier to read.

Also ignore how messy BuildableNodes.swift looks by now. I will try to clean it up once generate-swiftsyntaxbuilder has been migrated to use the parser and is thus easier to work with.