Decontextualize multi-'yield' Parsing
Yield statements support the ability to yield multiple values out of the coroutine context. These are modeled like an argument list in the grammar and parse sort of like one (minus labels). But this means there's a pretty serious grammatical ambiguity with applies

func yield() {}

yield ()

Rather than contextualize the parse, we're going to ignore the contextuality of multi-yields. The swift parser will always unambiguously parse 'yield &' as a yield statement, regardless of context. It will similarly unambiguously parse 'yield (' as an apply of a function called 'yield' regardless of context. We will then rewrite this AST on the way out to ASTGen where we actually have the ability to make context-sensitive decisions about the parse.