Use indentation to judge whether a closing `}` should be considered part of a code block if `{` is missing
## Motivating example:

```swift
struct Foo {
  func foo()
}
```

We need to decide whether we should use the `}` to close the opening `{` of the `struct` or `func`.

## Solution

Do what a human would do: Use to decide how to match the braces. If the opening `{` of the `func` is missing and the `}` has an indentation that’s less than that of the `func` keyword, assume that it was indented to match an outer and don’t eat it for the `func`.