Remove `print-init` subcommand from `swift-parser-cli`
`debugInitCall`, which is the backbone of `swift-parser-cli` is defined in `_SwiftSyntaxTestSupport`, which links against XCTest. If `swift-parser-cli` links against `_SwiftSyntaxTestSupport`, it also links against `XCTest`, causing its execution to fail with a dyld error outside of Xcode.

The solution would be to move `debugInitCall` into a new module that `swift-parser-cli` can link against but IMHO the `print-init` subcommand doesn’t provide sufficent value for that since you most likely want to use it in test cases where you can use a debugger command to get the `debugInitCall` result: `e print(tree.debugInitCall)`.