Allow leading dot syntax for StaticParserError
Convert StaticParserError into a struct, and
define the diagnostics in a
`DiagnosticMessage where Self == StaticParserError`
extension. This allows us to use leading dot syntax
to pass a StaticParserError anywhere a DiagnosticMessage
is expected.

To compute the diagnostic ID, we can (ab)use
`#function` to pick up the name of the property.