Eliminate PlatformMutex
Now that a new syntax arena is used for every mutation/node creation.
And the arena is never reused for allocation once the transaction
completes. So there's no chace that multiple threads allocates memory,
or adding children to an arena at the same time as long as the user
dealing with red-tree API only. That means mutex is now useless.

Of course there still be chances to cause data races when a user uses
SyntaxArena or RawSyntax directly. But it's now user's responsibility to
manage threads accordingly.