Improve builder string escape insertion
The `StringLiteralExpr.init(… content: …)` initializer was attempting to escape newlines, but it did not insert pound delimiters after the backslashes as Swift requires for raw literals. It also was not escaping control characters, which (except for tab) can only be written as escapes in Swift source code. Modify the initializer’s logic to correct these issues.