Lift @_spi To TokenConsumer
When Library Evolution is enabled, this protocol requires its @_spi members to have default implementations. For one, there's not really a reasonable default implementation we could provide. For another, there's not really a way one could conform to this protocol without also having SPI access to raw tokens from the lexer. Lift the SPI restriction to the entire protocol and refactor the conformances a bit.