[Macros] Remove the source location converter from MacroEvaluationContext.
Having the source location converter in the context exposes the entire
contents of the source file containing the macro expansion to the
implementation of the macro. While convenient for the macro author,
this breaks Swift's incremental build model, because Swift assumes
that (e.g.) changing the body of a function cannot affect the result of
type-checking the body of an unrelated function.

Drop the implementations of some "built-in" macros that are no longer
expressible, such as `#filePath`, `#line`, and `#column`.