[Macros] "Disconnect" macro expansion expression syntax from its parent.
Providing the macro expansion syntax node directly to the macro
implementation allowed those implementations to chase its parent
pointer, which indirectly exposed the full source code of the enclosing
file to the macro implementation. "Disconnect" the syntax node from
its parent so the syntax node is independent of all other source code
in the file.

Another sample macro implementation, `#function`, is now
unusable---but this is a better model for incremental builds.