Remove `TokenSyntax` shorthand static members for SwiftSyntaxBuilder
In an effort to unify SwiftSyntax and SwiftSyntaxBuilder I think we should remove the shorthand static members to create tokens in SwiftSyntaxBuilder that drop the `Token` suffix.

In a follow-up PR I will investigate whether it makes sense to drop the `Token` suffix from the token static members in SwiftSyntax.

rdar://104090850