Add bazel configuration
This allows bazel users to depend on these libraries directly through
the SwiftSyntax repo instead of everyone having to provide their own
duplicate BUILD files. Since swift-syntax is pretty uniform this is
setup using a simple bazel macro that uses the conventions of swiftpm to
setup each library. This also includes a bazel target for always
building a library with optimizations which can be useful if you want
the improved performance and aren't trying to debug swift-syntax itself.