Model `RawTokenKind` as a struct instead of an enum with an associated value for keywords
If we model `RawTokenKind` as an enum with associated values, equality comparisons call into `__derived_enum_equals`, which is less efficient than comparing the base and keyword of the `RawTokenKind` separately. I believe this is because `__derived_enum_equals` needs to handle the case where multiple cases have associated values, while we can special case for a single associated value.